EPITA/adrien.josse-42sh/tests/python/requirements.txt
argparse==1.4.0
pathlib2==2.3.5
six==1.13.0
termcolor==1.1.0
PyYAML==5.1.2
